nest game inside of set inside of match so parent properties can be accessed
